ci: add Ariane configuration file for 1.11
Ariane is a new GitHub App intended to trigger Cilium CI workflows based
on trigger phrases commented in pull requests, in order to replace the
existing `issue_comment`-based workflows and simplify our CI stack.

This commit adds a configuration setting up triggers such that existing
1.11 workflows can be triggered with the usual `/test-backport-1.11`,
and based on the same PR changelist match / ignore rules.

In particular for 1.11, the AKS workflow was NOT backported because 1.11
only supports K8s versions up to 1.23, and 1.23 is not available on AKS
anymore. The stable 1.11 AKS workflow from `main` was already disabled
when this changed happened, as per our testing policy, so we are not
backporting it.

Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>